# docs/swagger/swagger.yaml
definitions:
...
  model.Capture:
    properties:
      address:
        type: string
      cluster_id:
        type: string
      id:
        type: string
      is_owner:
        type: boolean
    type: object
...
  model.HTTPError:
    properties:
      error_code:
        type: string
      error_msg:
        type: string
    type: object
...
  v2.Capture:
    properties:
      address:
        type: string
      cluster_id:
        type: string
      id:
        type: string
      is_owner:
        type: boolean
    type: object
...
info:
  contact: {}
paths:
  /api/v1/captures:
    get:
      consumes:
      - application/json
      description: list all captures in cdc cluster
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Capture'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HTTPError'
      summary: List captures
      tags:
      - capture
...
  /api/v2/captures:
    get:
      description: list all captures in cdc cluster
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v2.Capture'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HTTPError'
      summary: List captures
      tags:
      - capture
      - v2
...
swagger: "2.0"